Added script mbhXmlFileFilter
Added a beta version of script mbhXmlFileFilter.py and its accompanying mbhXmlFileFilter.cms. This script filters an XML file by applying regular expressions, and optionally preforms search and replace on the filtered results. The results appear in a text file, that the user can optionally save as an XML file.